requirements.txt
fastapi
uvicorn[standard]
pymongo
redis
python-dotenv
